backend/pyproject.toml
[project]
name = "analyzr-backend"
version = "0.1.0"
description = "Enterprise-Grade Autonomous AI Data Analyst - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.7.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "statsmodels>=0.14.2",
    "scikit-learn>=1.5.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.2",
    "reportlab>=4.2.0",
    "python-docx>=1.1.2",
    "ollama>=0.2.1",
    "python-dotenv>=1.0.1",
    "aiofiles>=23.2.1",
    "httpx>=0.27.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
]
